Share KRX trading calendar in core for #47
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
